delete recruitment in profile
